debugging

+++ headneck/evaluate.py
@@ -13,7 +13,7 @@ from keras import backend as K
 
 def get_background_mask(data):
 
-    mask = np.array(data,dtype=np.uint8)
+    mask = np.array(data,dtype=np.float16)
     mask[data == 0] = 1
     mask[data == 1] = 0   
 
@@ -22,7 +22,7 @@ def get_background_mask(data):
 
 def get_organ_mask(data):
 
-    mask = np.array(data,dtype=np.uint8)
+    mask = np.array(data,dtype=np.float16)
 
     return mask
 
@@ -30,7 +30,7 @@ def dice_coefficient(y_true, y_pred, smooth=0.00001):
     y_true_f = K.flatten(y_true)
     y_pred_f = K.flatten(y_pred)
     intersection = K.sum(y_true_f * y_pred_f)
-    return (2 * intersection + smooth) / (K.sum(y_true_f) + K.sum(y_pred_f) + smooth)
+    return (2. * intersection + smooth) / (K.sum(y_true_f) + K.sum(y_pred_f) + smooth)
 
 #def dice_coefficient(truth, prediction):
 #    return 2 * np.sum(truth * prediction)/(np.sum(truth) + np.sum(prediction))
